+++ data/rules.lua
@@ -22,7 +22,7 @@ local modules = {
 };
 
 local result = {
-    version = 13,
+    version = 14,
     ability_rules = {},
     state_rules = {},
     catalog = {},